requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
openai>=1.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pandas>=2.0.0
numpy>=1.24.0
chromadb>=1.0.0
sqlglot>=26.0.0
streamlit>=1.38.0
streamlit-mermaid>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
